+++ src/UIState/SetKP.h
@@ -22,5 +22,5 @@ public:
   const char* prompt() {
     return "Set KP          ";
   };
-  void setValue(double value);
+  void setValue(float value);
 };
